# api/swagger.yml
swagger: "2.0"
info:
  description: "The ProjectJackson API. Azure-based services."
  version: "1.0.0"
  title: "Project Jackson API"
schemes:
- "https"
paths:
...
  "/people/{nconst}/titles":
    get:
      tags:
      - people
      - active
      summary: "Gets the titles associated with the person associated with specified ID"
      description: "Returns array of titles"
      produces:
      - "application/json"
      parameters:
      - in: path
        name: nconst
        description: "ID (nconst) of the person"
        required: true
        type: string
      responses:
        200:
          description: Successful return of titles
          schema:
            $ref: "#/definitions/Title"
        401:
          description: Unauthorized
        404:
          description: No valid ID Supplied
...
  "/titles/{tconst}/people":
    get:
      tags:
      - titles
      - active
      - principalwithname
      summary: "Gets the people popularly associated with the given title"
      description: "Returns list of PrincipalWithName objects"
      produces:
      - "application/json"
      parameters:
      - in: path
        name: tconst
        description: "tconst of title to get"
        required: true
        type: string
      responses:
        200:
          description: Successful get
          schema:
            $ref: "#/definitions/PrincipalWithName"
        401:
          description: Unauthorized
        404:
          description: Title not supplied/found
  "/titles/{tconst}/cast":
    get:
      tags:
      - titles
      - active
      - principalwithname
      summary: "Gets the actors associated with the given title"
      description: "Returns list of PrincipalWithName objects whose role is actor"
      produces:
      - "application/json"
      parameters:
      - in: path
        name: tconst
        description: "tconst of title to get"
        required: true
        type: string
      responses:
        200:
          description: Successful get
          schema:
            $ref: "#/definitions/PrincipalWithName"
        401:
          description: Unauthorized
        404:
          description: Title not supplied/found
  "/titles/{tconst}/crew":
    get:
      tags:
      - titles
      - active
      - principalwithname
      summary: "Gets the crew members associated with the given title"
      description: "Returns list of PrincipalWithName objects whose roles are not actor"
      produces:
      - "application/json"
      parameters:
      - in: path
        name: tconst
        description: "tconst of title to get"
        required: true
        type: string
      responses:
        200:
          description: Successful get
          schema:
            $ref: "#/definitions/PrincipalWithName"
        401:
          description: Unauthorized
        404:
          description: Title not supplied/found
...
definitions:
  Title:
    properties:
      tconst:
        description: "A string of a sequential number prepended with 'tt'"
        type: string
      titleType:
        description: "A string describing the type of content, eg. tvSeries, movie, tvMiniSeries"
        type: string
      primaryTitle:
        description: "A string of the title of the piece of media"
        type: string
      originalTitle:
        description: "A string of the original title of the piece of media, typically distinct from primaryTitle if film name is originally in a foreign language"
        type: string
      isAdult:
        description: "An integer representation of a boolean (0 or 1) whether the content is adult, ie. typically of pornographic nature"
        type: integer
        format: int32
      startYear:
        description: "An integer of the year; for non-serialized content it's the release year"
        type: integer
        format: int32
      endYear:
        description: "An integer of the year; for serialized content it's the year media completes"
        type: integer
        format: int32
      runtimeMinutes:
        description: "An integer of the number of minutes long the content lasts"
        type: integer
        format: int32
      genres:
        description: "A array of strings for genres associated with the piece"
        type: array
        items:
          type: string
...
  PrincipalWithName:
    properties:
      tconst:
        description: "A string representing a title's IMDb ID, one half of the title-person join"
        type: string
      ordering:
        description: "An integer representing the billing order when searching on IMDb for the piece of media"
        type: integer
        format: int32
      person:
        description: "An LinkedHashMap object that maps to the fields found in an a Person object"
        type: object
      category:
        description: "A string that contains what kind of role a Person has in a piece of media"
        type: integer
        format: int32
      job:
        description: "A string that has a person's job in the piece of media"
        type: string
      characters:
        description: "An array of strings that optionally contain what characters a Person plays in the piece of media"
        type: array
        items:
          type: string